Refactor to remove legacy solvers and rename solver_rewrite to solver.
