+++ vayesta/solver/_uccsd_eris.py
