+++ vayesta/solver/coupling.py
@@ -1,11 +1,10 @@
 import numpy as np
 import pyscf
-from vayesta.core.ao2mo import helper as ao2mo_helper
-from vayesta.core.util import *
+
 from vayesta.core import spinalg
+from vayesta.core.util import *
 from vayesta.mpi import mpi, RMA_Dict
-from vayesta.solver.simple import CCSD as SimpleCCSD
-from vayesta.solver_rewrite import ccsdtq
+from vayesta.solver import ccsdtq
 
 
 def transform_amplitude(t, u_occ, u_vir, u_occ2=None, u_vir2=None, inverse=False):
@@ -38,7 +37,7 @@ def transform_amplitude(t, u_occ, u_vir, u_occ2=None, u_vir2=None, inverse=False
         tab = transform_amplitude(t[1], u_occ[0], u_vir[0], u_occ[1], u_vir[1])
         tbb = transform_amplitude(t[2], u_occ[1], u_vir[1])
         return (taa, tab, tbb)
-    raise NotImplementedError("Transformation of %s amplitudes with ndim=%d" % (spinsym, np.ndim(t[0])+1))
+    raise NotImplementedError("Transformation of %s amplitudes with ndim=%d" % (spinsym, np.ndim(t[0]) + 1))
 
 
 def get_amplitude_norm(t1, t2):
@@ -48,8 +47,8 @@ def get_amplitude_norm(t1, t2):
         t2norm = np.linalg.norm(t2)
     # Unrestricted
     elif np.ndim(t1[0]) == 2:
-        t1norm = (np.linalg.norm(t1[0])+np.linalg.norm(t1[1]))/2
-        t2norm = (np.linalg.norm(t2[0])+2*np.linalg.norm(t2[1])+np.linalg.norm(t2[2]))/2
+        t1norm = (np.linalg.norm(t1[0]) + np.linalg.norm(t1[1])) / 2
+        t2norm = (np.linalg.norm(t2[0]) + 2 * np.linalg.norm(t2[1]) + np.linalg.norm(t2[2])) / 2
     return t1norm, t2norm
 
 
@@ -67,7 +66,7 @@ def project_t2_rspin(t2, proj, projectors):
         return t2
     if projectors == 1:
         t2 = einsum('xi,i...->x...', proj, t2)
-        return (t2 + t2.transpose(1,0,3,2))/2
+        return (t2 + t2.transpose(1, 0, 3, 2)) / 2
     if projectors == 2:
         return einsum('xi,yj,ij...->xy...', proj, proj, t2)
     raise ValueError
@@ -81,12 +80,12 @@ def project_t2_uspin(t2, proj, projectors):
     if projectors == 1:
         # Average between projecting alpha and beta:
         t2ab = (einsum('xi,ij...->xj...', proj[0], t2[1])
-              + einsum('xj,ij...->ix...', proj[1], t2[1]))/2
+                + einsum('xj,ij...->ix...', proj[1], t2[1])) / 2
     elif projectors == 2:
         t2ab = einsum('xi,yj,ij...->xy...', proj[0], proj[1], t2[1])
     else:
         raise ValueError
-    #assert np.allclose(t2ab, -t2ab.transpose(0,1,3,2))
+    # assert np.allclose(t2ab, -t2ab.transpose(0,1,3,2))
     return (t2aa, t2ab, t2bb)
 
 
@@ -136,10 +135,10 @@ def couple_ccsd_iterations(solver, fragments):
             po = p_occ[y.id]
             ro = r_occ[y.id]
             rv = r_vir[y.id]
-            #print(solver.fragment.id, y.id, py.shape, t1_out.shape, t1y.shape)
+            # print(solver.fragment.id, y.id, py.shape, t1_out.shape, t1y.shape)
             t1_out += einsum('Ii,ia,Aa->IA', po, t1y, rv)
             t2_out += einsum('Ii,Jj,ijab,Aa,Bb->IJAB', po, ro, t2y, rv, rv)
-        solver.log.info("Tailoring: |dT1|= %.3e  |dT2|= %.3e", np.linalg.norm(t1_out-t1), np.linalg.norm(t2_out-t2))
+        solver.log.info("Tailoring: |dT1|= %.3e  |dT2|= %.3e", np.linalg.norm(t1_out - t1), np.linalg.norm(t2_out - t2))
         rma.clear()
         t1[:] = t1_out
         t2[:] = t2_out
@@ -173,16 +172,19 @@ def tailor_with_fragments(solver, fragments, project=False, tailor_t1=True, tail
     tailor_func : function(cc, t1, t2) -> t1, t2
         Tailoring function for CCSD.
     """
-    fragment = solver.fragment
-    cluster = solver.cluster
-    ovlp = solver.base.get_ovlp()       # AO overlap matrix
-    cx_occ = cluster.c_active_occ       # Occupied active orbitals of current cluster
-    cx_vir = cluster.c_active_vir       # Virtual  active orbitals of current cluster
+    fragment = solver.hamil._fragment
+    cluster = solver.hamil.cluster
+    base = fragment.base
+    ovlp = base.get_ovlp()  # AO overlap matrix
+    cx_occ = cluster.c_active_occ  # Occupied active orbitals of current cluster
+    cx_vir = cluster.c_active_vir  # Virtual  active orbitals of current cluster
     cxs_occ = spinalg.dot(spinalg.T(cx_occ), ovlp)
     cxs_vir = spinalg.dot(spinalg.T(cx_vir), ovlp)
     project = int(project)
-    nxy_occ = solver.base.get_fragment_overlap_norm(fragments=([fragment], fragments), virtual=False, norm=None)[0]
-    nxy_vir = solver.base.get_fragment_overlap_norm(fragments=([fragment], fragments), occupied=False, norm=None)[0]
+    nxy_occ = fragment.base.get_fragment_overlap_norm(fragments=([fragment], fragments), virtual=False, norm=None)[0]
+    nxy_vir = fragment.base.get_fragment_overlap_norm(fragments=([fragment], fragments), occupied=False, norm=None)[0]
+
+    spinsym = base.spinsym
 
     def tailor_func(kwargs):
         """Add external correction to T1 and T2 amplitudes."""
@@ -201,13 +203,13 @@ def tailor_with_fragments(solver, fragments, project=False, tailor_t1=True, tail
             rxy_occ = spinalg.dot(cxs_occ, fy.cluster.c_active_occ)
             rxy_vir = spinalg.dot(cxs_vir, fy.cluster.c_active_vir)
             # Skip fragment if there is no overlap
-            if solver.spinsym == 'restricted':
+            if spinsym == 'restricted':
                 maxovlp = min(abs(rxy_occ).max(), abs(rxy_vir).max())
-            elif solver.spinsym == 'unrestricted':
+            elif spinsym == 'unrestricted':
                 maxovlp = min(max(abs(rxy_occ[0]).max(), abs(rxy_occ[1]).max()),
                               max(abs(rxy_vir[0]).max(), abs(rxy_vir[1]).max()))
             if maxovlp < ovlp_tol:
-                self.log.debug("Skipping tailoring fragment %s due to small overlap= %.1e", fy, maxovlp)
+                solver.log.debug("Skipping tailoring fragment %s due to small overlap= %.1e", fy, maxovlp)
                 continue
 
             wfy = fy.results.wf.as_ccsd()
@@ -240,15 +242,15 @@ def tailor_with_fragments(solver, fragments, project=False, tailor_t1=True, tail
 
         # Add correction:
         if tailor_t1:
-            if solver.spinsym == 'restricted':
+            if spinsym == 'restricted':
                 t1[:] += dt1
-            elif solver.spinsym == 'unrestricted':
+            elif spinsym == 'unrestricted':
                 t1[0][:] += dt1[0]
                 t1[1][:] += dt1[1]
         if tailor_t2:
-            if solver.spinsym == 'restricted':
+            if spinsym == 'restricted':
                 t2[:] += dt2
-            elif solver.spinsym == 'unrestricted':
+            elif spinsym == 'unrestricted':
                 t2[0][:] += dt2[0]
                 t2[1][:] += dt2[1]
                 t2[2][:] += dt2[2]
@@ -258,22 +260,17 @@ def tailor_with_fragments(solver, fragments, project=False, tailor_t1=True, tail
 
 
 def _integrals_for_extcorr(fragment, fock):
-    eris = fragment._eris
     cluster = fragment.cluster
     emb = fragment.base
-    if eris is None:
-        if emb.spinsym == 'restricted':
-            eris = emb.get_eris_array(cluster.c_active)
-        else:
-            eris = emb.get_eris_array_uhf(cluster.c_active)
+    eris = fragment.hamil.get_eris_screened()
 
     if emb.spinsym == 'restricted':
         occ = np.s_[:cluster.nocc_active]
         vir = np.s_[cluster.nocc_active:]
-        govov = eris[occ,vir,occ,vir] # chemical notation 
-        gvvov = eris[vir,vir,occ,vir]
-        gooov = eris[occ,occ,occ,vir]
-        govoo = eris[occ,vir,occ,occ]
+        govov = eris[occ, vir, occ, vir]  # chemical notation
+        gvvov = eris[vir, vir, occ, vir]
+        gooov = eris[occ, occ, occ, vir]
+        govoo = eris[occ, vir, occ, occ]
         fov = dot(cluster.c_active_occ.T, fock, cluster.c_active_vir)
         return fov, (govov, gvvov, gooov, govoo)
 
@@ -285,15 +282,16 @@ def _integrals_for_extcorr(fragment, fock):
         fova = dot(cluster.c_active_occ[0].T, fock[0], cluster.c_active_vir[0])
         fovb = dot(cluster.c_active_occ[1].T, fock[1], cluster.c_active_vir[1])
         fov = (fova, fovb)
-        gooov = (eris[0][oa,oa,oa,va], eris[1][oa,oa,ob,vb], eris[2][ob,ob,ob,vb])
-        govov = (eris[0][oa,va,oa,va], eris[1][oa,va,ob,vb], eris[2][ob,vb,ob,vb])
-        govvv = (eris[0][oa,va,va,va], eris[1][oa,va,vb,vb], eris[2][ob,vb,vb,vb])
-        gvvov = (None, eris[1][va,va,ob,vb], None)
-        govoo = (None, eris[1][oa,va,ob,ob], None)
+        gooov = (eris[0][oa, oa, oa, va], eris[1][oa, oa, ob, vb], eris[2][ob, ob, ob, vb])
+        govov = (eris[0][oa, va, oa, va], eris[1][oa, va, ob, vb], eris[2][ob, vb, ob, vb])
+        govvv = (eris[0][oa, va, va, va], eris[1][oa, va, vb, vb], eris[2][ob, vb, vb, vb])
+        gvvov = (None, eris[1][va, va, ob, vb], None)
+        govoo = (None, eris[1][oa, va, ob, ob], None)
         return fov, (gooov, govov, govvv, gvvov, govoo)
 
     else:
         raise NotImplementedError(emb.spinsym)
+
 
 def _get_delta_t_for_extcorr(fragment, fock, solver, include_t3v=True):
     """Make T3 and T4 residual correction to CCSD wave function for given fragment.
@@ -330,7 +328,7 @@ def _get_delta_t_for_extcorr(fragment, fock, solver, include_t3v=True):
     # --- Make correction to T1 and T2 amplitudes
     # J. Chem. Theory Comput. 2021, 17, 182−190
     # also with reference to git@github.com:gustavojra/Methods.git
-    
+
     if fragment.base.spinsym == 'restricted':
         # Get ERIs and Fock matrix for the given fragment
         f, v = _integrals_for_extcorr(fragment, fock)
@@ -349,6 +347,7 @@ def _get_delta_t_for_extcorr(fragment, fock, solver, include_t3v=True):
         raise ValueError
 
     return dt1, dt2
+
 
 def _get_delta_t2_from_t3v(govvv_x, gvvov_x, gooov_x, govoo_x, frag_child, rxy_occ, rxy_vir, cxs_occ, projectors):
     """Perform the (T3 * V) contraction for the external correction, with the V integrals
@@ -374,7 +373,7 @@ def _get_delta_t2_from_t3v(govvv_x, gvvov_x, gooov_x, govoo_x, frag_child, rxy_o
     dt2 : numpy.ndarray
         Update to T2 amplitudes in the parent (x) basis
     """
-    
+
     wf = frag_child.results.wf.as_ccsdtq()
     t3 = wf.t3
 
@@ -386,7 +385,7 @@ def _get_delta_t2_from_t3v(govvv_x, gvvov_x, gooov_x, govoo_x, frag_child, rxy_o
         gvvov_ = einsum('abic,bB,iI,cC -> aBIC', gvvov_x, rxy_vir, rxy_occ, rxy_vir)
 
         # Contract with T3 amplitudes in the y basis
-        t3v_ = 0.5*einsum('bemf, jimeaf -> ijab', gvvov_ - gvvov_.transpose(0,3,2,1), t3)
+        t3v_ = 0.5 * einsum('bemf, jimeaf -> ijab', gvvov_ - gvvov_.transpose(0, 3, 2, 1), t3)
         t3v_ += einsum('bemf, ijmaef -> ijab', gvvov_, t3)
         # Final is in a mixed basis form, with last index in t3v here in the x basis
         # Rotate remaining indices into x basis: another three-quarter transform
@@ -399,13 +398,13 @@ def _get_delta_t2_from_t3v(govvv_x, gvvov_x, gooov_x, govoo_x, frag_child, rxy_o
         govoo_ = einsum('iajk,iI,aA,jJ -> IAJk', govoo_x, rxy_occ, rxy_vir, rxy_occ)
 
         # Second index of t3v_ in the parent (x) basis
-        t3v_ = -0.5*einsum('mjne, minbae -> ijab', gooov_ - govoo_.transpose(0,3,2,1), t3)
+        t3v_ = -0.5 * einsum('mjne, minbae -> ijab', gooov_ - govoo_.transpose(0, 3, 2, 1), t3)
         t3v_ -= einsum('mjne, imnabe -> ijab', gooov_, t3)
         # Rotate remaining indices into x basis: another three-quarter transform
         t3v_x += einsum('IjAB,iI,aA,bB -> ijab', t3v_, rxy_occ, rxy_vir, rxy_vir)
 
         # Include permutation
-        dt2 = t3v_x + t3v_x.transpose(1,0,3,2)
+        dt2 = t3v_x + t3v_x.transpose(1, 0, 3, 2)
 
     elif frag_child.base.spinsym == 'unrestricted':
         gooov_aaaa, gooov_aabb, gooov_bbbb = gooov_x
@@ -433,49 +432,49 @@ def _get_delta_t2_from_t3v(govvv_x, gvvov_x, gooov_x, govoo_x, frag_child, rxy_o
 
         x0 = einsum("Jlkc,iklacb->iJab", gooov_aabb_, t3_aba)
         x1 = einsum("Jlkc,iklabc->iJab", gooov_aaaa_, t3_aaa) * -1.0
-        x2 =  einsum("iJba->iJab", x0) * -1.0
+        x2 = einsum("iJba->iJab", x0) * -1.0
         x2 += einsum("iJba->iJab", x1) * -1.0
-        x2 =  einsum("iJab,Ii,Aa,Bb->IJAB", x2, rxy_occ[0], rxy_vir[0], rxy_vir[0])
-        dt2_aaaa =  einsum("ijab->ijab", x2) * -1.0
+        x2 = einsum("iJab,Ii,Aa,Bb->IJAB", x2, rxy_occ[0], rxy_vir[0], rxy_vir[0])
+        dt2_aaaa = einsum("ijab->ijab", x2) * -1.0
         dt2_aaaa += einsum("jiab->ijab", x2)
 
         x0 = einsum("Bdkc,ikjacd->ijaB", gvvov_aabb_, t3_aba)
         x1 = einsum("kcBd,ijkacd->ijaB", govvv_aaaa_, t3_aaa)
-        x2 =  einsum("ijaB->ijaB", x0)
+        x2 = einsum("ijaB->ijaB", x0)
         x2 += einsum("ijaB->ijaB", x1) * -1.0
-        x2 =  einsum("ijaB,Ii,Jj,Aa->IJAB", x2, rxy_occ[0], rxy_occ[0], rxy_vir[0])
+        x2 = einsum("ijaB,Ii,Jj,Aa->IJAB", x2, rxy_occ[0], rxy_occ[0], rxy_vir[0])
         dt2_aaaa += einsum("ijab->ijab", x2)
         dt2_aaaa += einsum("ijba->ijab", x2) * -1.0
 
         x0 = einsum("Jklc,iklabc->iJab", gooov_bbbb_, t3_bbb)
         x1 = einsum("lcJk,ilkacb->iJab", govoo_aabb_, t3_bab)
-        x2 =  einsum("iJba->iJab", x0) * -1.0
+        x2 = einsum("iJba->iJab", x0) * -1.0
         x2 += einsum("iJba->iJab", x1) * -1.0
-        x2 =  einsum("iJab,Ii,Aa,Bb->IJAB", x2, rxy_occ[1], rxy_vir[1], rxy_vir[1])
-        dt2_bbbb =  einsum("ijab->ijab", x2) * -1.0
+        x2 = einsum("iJab,Ii,Aa,Bb->IJAB", x2, rxy_occ[1], rxy_vir[1], rxy_vir[1])
+        dt2_bbbb = einsum("ijab->ijab", x2) * -1.0
         dt2_bbbb += einsum("jiab->ijab", x2)
 
         x0 = einsum("kdBc,ijkacd->ijaB", govvv_bbbb_, t3_bbb) * -1.0
         x1 = einsum("kdBc,ikjadc->ijaB", govvv_aabb_, t3_bab)
-        x2 =  einsum("ijaB->ijaB", x0)
+        x2 = einsum("ijaB->ijaB", x0)
         x2 += einsum("ijaB->ijaB", x1) * -1.0
-        x2 =  einsum("ijaB,Ii,Jj,Aa->IJAB", x2, rxy_occ[1], rxy_occ[1], rxy_vir[1])
+        x2 = einsum("ijaB,Ii,Jj,Aa->IJAB", x2, rxy_occ[1], rxy_occ[1], rxy_vir[1])
         dt2_bbbb += einsum("ijab->ijab", x2) * -1.0
         dt2_bbbb += einsum("ijba->ijab", x2)
 
-        x0 =  einsum("Ilkc,jlkbac->Ijab", gooov_aabb_, t3_bab) * -1.0
+        x0 = einsum("Ilkc,jlkbac->Ijab", gooov_aabb_, t3_bab) * -1.0
         x0 += einsum("Ilmd,ljmabd->Ijab", gooov_aaaa_, t3_aba) * -1.0
-        dt2_abab =  einsum("Ijab,Jj,Aa,Bb->IJAB", x0, rxy_occ[1], rxy_vir[0], rxy_vir[1])
+        dt2_abab = einsum("Ijab,Jj,Aa,Bb->IJAB", x0, rxy_occ[1], rxy_vir[0], rxy_vir[1])
 
-        x0 =  einsum("Jnkc,kinbac->iJab", gooov_bbbb_, t3_bab)
+        x0 = einsum("Jnkc,kinbac->iJab", gooov_bbbb_, t3_bab)
         x0 += einsum("ldJk,iklabd->iJab", govoo_aabb_, t3_aba) * -1.0
         dt2_abab += einsum("iJab,Ii,Aa,Bb->IJAB", x0, rxy_occ[0], rxy_vir[0], rxy_vir[1])
 
-        x0 =  einsum("ldAe,ijldbe->ijAb", govvv_aaaa_, t3_aba) * -1.0
+        x0 = einsum("ldAe,ijldbe->ijAb", govvv_aaaa_, t3_aba) * -1.0
         x0 += einsum("Adkc,jikbdc->ijAb", gvvov_aabb_, t3_bab)
         dt2_abab += einsum("ijAb,Ii,Jj,Bb->IJAB", x0, rxy_occ[0], rxy_occ[1], rxy_vir[1])
 
-        x0 =  einsum("ldBc,ijlacd->ijaB", govvv_aabb_, t3_aba)
+        x0 = einsum("ldBc,ijlacd->ijaB", govvv_aabb_, t3_aba)
         x0 += einsum("kcBf,jikcaf->ijaB", govvv_bbbb_, t3_bab) * -1.0
         dt2_abab += einsum("ijaB,Ii,Jj,Aa->IJAB", x0, rxy_occ[0], rxy_occ[1], rxy_vir[0])
 
@@ -487,46 +486,29 @@ def _get_delta_t2_from_t3v(govvv_x, gvvov_x, gooov_x, govoo_x, frag_child, rxy_o
     # Find the fragment projector of cluster y (child) in the basis of cluster x (parent)
     c_frag_xocc = spinalg.dot(spinalg.T(frag_child.c_frag), spinalg.T(cxs_occ))
     proj_y_in_x = spinalg.dot(spinalg.T(c_frag_xocc), c_frag_xocc)
-    
+
     # Project (t3 v) contribution onto fragment of cluster y
     dt2 = project_t2(dt2, proj_y_in_x, projectors=projectors)
 
     return dt2
 
-def _get_delta_t_for_delta_tailor(fragment, fock):
+
+def _get_delta_t_for_delta_tailor(fragment):
     wf = fragment.results.wf.as_ccsd()
     t1, t2 = wf.t1, wf.t2
-    # CCSD
-    cluster = fragment.cluster
-    fock = spinalg.dot(spinalg.T(cluster.c_active), fock, cluster.c_active)
-    nocc = wf.mo.nocc
-    if fragment.base.spinsym == 'restricted':
-        mo_energy = np.diag(fock).copy()
-        if fragment.base.has_exxdiv:
-            mo_energy[:nocc] -= fragment.base.madelung
-    else:
-        mo_energy = (np.diag(fock[0]).copy(), np.diag(fock[1]).copy())
-        if fragment.base.has_exxdiv:
-            moa, mob = mo_energy
-            moa[:nocc[0]] -= fragment.base.madelung
-            mob[:nocc[1]] -= fragment.base.madelung
-            mo_energy = (moa, mob)
-
-    eris = fragment._eris
-    if eris is None:
-        # Symmetry-derived fragments may not have eris stored
-        # TODO: This may not scale well, since we are performing
-        # integral transformations for all fragments. Can be improved.
-        eris = fragment.base.get_eris_array(cluster.c_active)
-    ccsd = SimpleCCSD(fock, eris, nocc, mo_energy=mo_energy)
-    wf = ccsd.kernel(t1=t1, t2=t2)
+    # Run CCSD calculation with same Hamiltonian. Don't need lambda amplitudes.
+    # TODO set up passing through solver_options from tailored CCSD calculation
+    ccsd = fragment.get_solver("CCSD")
+    ccsd.opts.solve_lambda = False
+    ccsd.kernel()
     assert ccsd.converged
+    wf = ccsd.wf
     dt1 = spinalg.subtract(t1, wf.t1)
     dt2 = spinalg.subtract(t2, wf.t2)
     return dt1, dt2
 
 
-def externally_correct(solver, external_corrections, eris=None):
+def externally_correct(solver, external_corrections, hamil=None):  # eris=None):
     """Build callback function for CCSD, to add external correction from other fragments.
 
     TODO: combine with `tailor_with_fragments`?
@@ -542,7 +524,7 @@ def externally_correct(solver, external_corrections, eris=None):
     eris : _ChemistsERIs
         ERIs for parent CCSD fragment. Used for MO energies in residual contraction, and for
         the case of low_level_coul, where the parent Coulomb integral is contracted.
-        If not passed in, MO energy if needed will be constructed from the diagonal of 
+        If not passed in, MO energy if needed will be constructed from the diagonal of
         get_fock() of embedding base class, and the eris will be also be obtained from the
         embedding base class. Optional.
 
@@ -552,23 +534,23 @@ def externally_correct(solver, external_corrections, eris=None):
         Callback function for PySCF's CCSD solver.
     """
 
-    fx = solver.fragment
-    cluster = solver.cluster
-    emb = solver.base
+    fx = solver.hamil._fragment
+    cluster = solver.hamil.cluster
+    emb = fx.base
     nocc = cluster.nocc
     nvir = cluster.nvir
-    ovlp = emb.get_ovlp()               # AO overlap matrix
-    cx_occ = cluster.c_active_occ       # Occupied active orbitals of current cluster
-    cx_vir = cluster.c_active_vir       # Virtual  active orbitals of current cluster
+    ovlp = emb.get_ovlp()  # AO overlap matrix
+    cx_occ = cluster.c_active_occ  # Occupied active orbitals of current cluster
+    cx_vir = cluster.c_active_vir  # Virtual  active orbitals of current cluster
     cxs_occ = spinalg.dot(spinalg.T(cx_occ), ovlp)
     cxs_vir = spinalg.dot(spinalg.T(cx_vir), ovlp)
-    if eris is None:
+    if hamil is None:
         # Note that if no MO energies are passed in, we construct them from the 
         # get_fock function without with_exxdiv=False. For PBC CCSD, this may be different
         # behaviour.
         mo_energy = einsum('ai,ab,bi->i', cluster.c_active, emb.get_fock(), cluster.c_active)
     else:
-        mo_energy = eris.mo_energy
+        mo_energy = hamil.get_clus_mf_info(with_exxdiv=False)[2]  # Do we want this True or False?
 
     if (len(external_corrections) > 1) and any([corr[2] == 0 for corr in external_corrections]):
         # We are externally correcting from multiple fragments, but not projecting them
@@ -586,33 +568,27 @@ def externally_correct(solver, external_corrections, eris=None):
         # integrals in the parent (i.e. CCSD) cluster. We can take the integrals from eris
         # if passed in. Otherwise, form the required integrals
         # for this parent cluster. Note that not all of these are needed.
-        if eris is None:
+        if hamil is None:
             _, govvv_x, gvvov_x, gooov_x, govoo_x = _integrals_for_extcorr(fx, fock)
         else:
+            eri_generator = hamil.get_eris_bare
+
             if emb.spinsym == 'restricted':
                 govvv_x = None
-                gvvov_x = np.array(ao2mo_helper.get_ovvv(eris)).transpose(2,3,0,1)
-                gooov_x = np.array(eris.ovoo).transpose(2,3,0,1)
-                govoo_x = np.array(eris.ovoo)
-            elif emb.spinsym == 'unrestricted':
-                govvv_x = (np.array(ao2mo_helper.get_ovvv(eris, block="ovvv")),
-                           np.array(ao2mo_helper.get_ovvv(eris, block="ovVV")),
-                           np.array(ao2mo_helper.get_ovvv(eris, block="OVVV")))
-                gvvov_x = (np.array(ao2mo_helper.get_ovvv(eris, block="ovvv")).transpose(2,3,0,1),
-                           np.array(ao2mo_helper.get_ovvv(eris, block="OVvv")).transpose(2,3,0,1),
-                           np.array(ao2mo_helper.get_ovvv(eris, block="OVVV")).transpose(2,3,0,1))
-                gooov_x = (np.array(eris.ovoo).transpose(2,3,0,1),
-                           np.array(eris.OVoo).transpose(2,3,0,1),
-                           np.array(eris.OVOO).transpose(2,3,0,1))
-                govoo_x = (np.array(eris.ovoo),
-                           np.array(eris.ovOO),
-                           np.array(eris.OVOO))
-    
+                gvvov_x = eri_generator("vvov")
+                gooov_x = eri_generator("ooov")
+                govoo_x = eri_generator("ovoo")
+            else:
+                govvv_x = (eri_generator("ovvv"), eri_generator("ovVV"), eri_generator("OVVV"))
+                gvvov_x = (eri_generator("vvov"), eri_generator("vvOV"), eri_generator("VVOV"))
+                gooov_x = (eri_generator("ooov"), eri_generator("ooOV"), eri_generator("OOOV"))
+                govoo_x = (eri_generator("ovoo"), eri_generator("ovOO"), eri_generator("OVOO"))
+
     # delta-T1 and delta-T2 amplitudes, to be added to the CCSD amplitudes
-    if solver.spinsym == 'restricted':
+    if emb.spinsym == 'restricted':
         dt1 = np.zeros((nocc, nvir))
         dt2 = np.zeros((nocc, nocc, nvir, nvir))
-    elif solver.spinsym == 'unrestricted':
+    elif emb.spinsym == 'unrestricted':
         dt1 = (np.zeros((nocc[0], nvir[0])),
                np.zeros((nocc[1], nvir[1])))
         dt2 = (np.zeros((nocc[0], nocc[0], nvir[0], nvir[0])),
@@ -622,7 +598,7 @@ def externally_correct(solver, external_corrections, eris=None):
     frag_dir = {f.id: f for f in emb.fragments}
     for y, corrtype, projectors, low_level_coul in external_corrections:
 
-        fy = frag_dir[y] # Get fragment y object from its index
+        fy = frag_dir[y]  # Get fragment y object from its index
         assert (y != fx.id)
 
         if corrtype == 'external':
@@ -631,7 +607,7 @@ def externally_correct(solver, external_corrections, eris=None):
             else:
                 dt1y, dt2y = _get_delta_t_for_extcorr(fy, fock, solver, include_t3v=True)
         elif corrtype == 'delta-tailor':
-            dt1y, dt2y = _get_delta_t_for_delta_tailor(fy, fock)
+            dt1y, dt2y = _get_delta_t_for_delta_tailor(fy)
         else:
             raise ValueError
 
@@ -652,7 +628,7 @@ def externally_correct(solver, external_corrections, eris=None):
         dt1 = spinalg.add(dt1, dt1y)
         dt2 = spinalg.add(dt2, dt2y)
 
-        if low_level_coul and corrtype == 'external': 
+        if low_level_coul and corrtype == 'external':
             # Include the t3v term, contracting with the integrals from the x cluster
             # These have already been fragment projected, and rotated into the x cluster
             # in this function.
@@ -662,9 +638,9 @@ def externally_correct(solver, external_corrections, eris=None):
 
         solver.log.info("External correction residuals from fragment %3d (%s via %s):  dT1= %.3e  dT2= %.3e",
                         fy.id, fy.solver, corrtype, *get_amplitude_norm(dt1y, dt2y))
-    
-    if solver.spinsym == 'restricted':
-        
+
+    if emb.spinsym == 'restricted':
+
         if corrtype == 'external':
             # Contract with fragment x (CCSD) energy denominators
             # Note that this will not work correctly if a level shift used
@@ -674,7 +650,7 @@ def externally_correct(solver, external_corrections, eris=None):
             dt2 /= eijab
 
         solver.log.info("Total external correction amplitudes from all fragments:  dT1= %.3e  dT2= %.3e",
-                *get_amplitude_norm(dt1, dt2))
+                        *get_amplitude_norm(dt1, dt2))
 
         def callback(kwargs):
             """Add external correction to T1 and T2 amplitudes."""
@@ -682,7 +658,7 @@ def externally_correct(solver, external_corrections, eris=None):
             t1[:] += dt1
             t2[:] += dt2
 
-    elif solver.spinsym == 'unrestricted':
+    elif emb.spinsym == 'unrestricted':
 
         if corrtype == "external":
             eia_a = mo_energy[0][:nocc[0], None] - mo_energy[0][None, nocc[0]:]
@@ -695,7 +671,7 @@ def externally_correct(solver, external_corrections, eris=None):
             dt2 = (dt2[0] / eijab_aa, dt2[1] / eijab_ab, dt2[2] / eijab_bb)
 
         solver.log.info("Total external correction amplitudes from all fragments:  dT1= %.3e  dT2= %.3e", \
-                *get_amplitude_norm(dt1, dt2))
+                        *get_amplitude_norm(dt1, dt2))
 
         def callback(kwargs):
             """Add external correction to T1 and T2 amplitudes."""
