+++ vayesta/solver/dump.py
@@ -1,26 +1,24 @@
 import dataclasses
+
 import h5py
-import vayesta
+
 from vayesta.core import spinalg
 from .solver import ClusterSolver
 
 
 class DumpSolver(ClusterSolver):
-
     @dataclasses.dataclass
     class Options(ClusterSolver.Options):
-        filename: str = None
+        dumpfile: str = None
 
-    def kernel(self, *args, eris=None, **kwargs):
-        fragment = self.fragment
-        cluster = self.cluster
+    def kernel(self, *args, **kwargs):
+        fragment = self.hamil._fragment
+        cluster = self.hamil.cluster
 
-        if isinstance(self.opts.filename, str):
-            h5file = h5py.File(self.opts.filename, 'a')
+        if isinstance(self.opts.dumpfile, str):
+            h5file = h5py.File(self.opts.dumpfile, 'a')
         else:
-            h5file = self.opts.filename
-        if eris is None:
-            eris = self.get_eris()
+            h5file = self.opts.dumpfile
         with h5file as f:
             grp = h5file.create_group('fragment_%d' % fragment.id)
             # Attributes
@@ -32,7 +30,8 @@ class DumpSolver(ClusterSolver):
             c_dmet_cluster_occ = fragment._dmet_bath.c_cluster_occ
             c_dmet_cluster_vir = fragment._dmet_bath.c_cluster_vir
             c_dmet_cluster = spinalg.hstack_matrices(c_dmet_cluster_occ, c_dmet_cluster_vir)
-            if self.spinsym == 'restricted':
+            heff, eris = self.hamil.get_integrals(with_vext=True)
+            if c_dmet_cluster[0].ndim == 1:
                 grp.attrs['norb_dmet_cluster'] = c_dmet_cluster.shape[-1]
                 grp.attrs['nocc_dmet_cluster'] = c_dmet_cluster_occ.shape[-1]
                 grp.attrs['nvir_dmet_cluster'] = c_dmet_cluster_vir.shape[-1]
@@ -41,11 +40,10 @@ class DumpSolver(ClusterSolver):
                 grp.create_dataset('c_dmet_cluster', data=c_dmet_cluster)
                 grp.create_dataset('c_cluster', data=cluster.c_active)
                 # Integrals
-                grp.create_dataset('hcore', data=self.get_hcore())
-                grp.create_dataset('heff', data=self.get_heff(eris=eris))
-                grp.create_dataset('fock', data=self.get_fock())
+                grp.create_dataset('heff', data=heff)
+                grp.create_dataset('fock', data=self.hamil.get_fock())
                 grp.create_dataset('eris', data=eris)
-            elif self.spinsym == 'unrestricted':
+            elif c_dmet_cluster[0].ndim == 2:
                 grp.attrs['norb_dmet_cluster'] = [c_dmet_cluster[s].shape[-1] for s in range(2)]
                 grp.attrs['nocc_dmet_cluster'] = [c_dmet_cluster_occ[s].shape[-1] for s in range(2)]
                 grp.attrs['nvir_dmet_cluster'] = [c_dmet_cluster_vir[s].shape[-1] for s in range(2)]
@@ -57,13 +55,10 @@ class DumpSolver(ClusterSolver):
                 grp.create_dataset('c_cluster_a', data=cluster.c_active[0])
                 grp.create_dataset('c_cluster_b', data=cluster.c_active[1])
                 # Integrals
-                hcorea, hcoreb = self.get_hcore()
-                grp.create_dataset('hcore_a', data=hcorea)
-                grp.create_dataset('hcore_b', data=hcoreb)
-                heffa, heffb = self.get_heff(eris=eris)
+                heffa, heffb = heff
                 grp.create_dataset('heff_a', data=heffa)
                 grp.create_dataset('heff_b', data=heffb)
-                focka, fockb = self.get_fock()
+                focka, fockb = self.hamil.get_fock()
                 grp.create_dataset('fock_a', data=focka)
                 grp.create_dataset('fock_b', data=fockb)
                 erisaa, erisab, erisbb = eris
