--- vayesta/solver/simple/__init__.py
@@ -1,3 +1,0 @@
-from .ccsd import CCSD
-from .ccsd import RCCSD
-from .ccsd import UCCSD
